requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
Requests==2.31.0
telebot~=0.0.5